Methods/ragQA/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
tiktoken>=0.5.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=0.21.0
